+++ src/pari.cpp
@@ -5,8 +5,6 @@
 
 #include "log.hpp"
 
-std::string Pari::generate(const Formula& f) { return f.toString(); }
-
 Sequence Pari::eval(const Formula& f, int64_t start, int64_t end) {
   const std::string gpPath("pari-loda.gp");
   const std::string gpResult("pari-result.txt");
@@ -14,7 +12,7 @@ Sequence Pari::eval(const Formula& f, int64_t start, int64_t end) {
   if (!gp) {
     throw std::runtime_error("error generating gp file");
   }
-  gp << generate(f) << std::endl;
+  gp << f.toString(true) << std::endl;
   gp << "for (n = " << start << ", " << end << ", print(a(n)))" << std::endl;
   gp << "quit" << std::endl;
   gp.close();